debug and rebug with docker

+++ get_next_line.h
@@ -6,7 +6,7 @@
 /*   By: smetzler <smetzler@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/08/25 09:42:31 by smetzler          #+#    #+#             */
-/*   Updated: 2021/09/15 16:58:41 by smetzler         ###   ########.fr       */
+/*   Updated: 2021/09/16 18:32:59 by smetzler         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@
 
 char	*get_next_line(int fd);
 char	*ft_makereturn(char **tonext, int location);
-char	*ft_calloc(int size, char filler);
+char	*ft_calloc(size_t size, char filler);
 char	*ft_prepnext(char **tonext, int location, int size);
 void	ft_free(char **line);
 int		ft_strlen(char *str);
